update image so it can handle files greatter than 2 Gb
git-svn-id: file:///home/roberto/Git/svn@6526 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: ccbc0c36bb9cee0aba55f75f5398d16a701c7db2

+++ trunk/xmipp/libraries/data/metadata_extension.cpp
@@ -72,14 +72,14 @@ void getStatistics(MetaData &MT_in, Image<double> & _ave, Image<double> & _sd, d
     _sd().selfSQRT();
 }
 
-void SingleImgSize(const FileName &filename, int &Xdim, int &Ydim, int &Zdim, int &Ndim)
+void SingleImgSize(const FileName &filename, int &Xdim, int &Ydim, int &Zdim, unsigned long &Ndim)
 {
     Image<double> img;
     img.read(filename, false);
     img.getDimensions(Xdim, Ydim, Zdim, Ndim);
 }
 
-void ImgSize(const MetaData &MD, int &Xdim, int &Ydim, int &Zdim, int &Ndim)
+void ImgSize(const MetaData &MD, int &Xdim, int &Ydim, int &Zdim, unsigned long &Ndim)
 {
     if (!MD.isEmpty())
     {
@@ -92,7 +92,7 @@ void ImgSize(const MetaData &MD, int &Xdim, int &Ydim, int &Zdim, int &Ndim)
         REPORT_ERROR(ERR_MD_NOOBJ, "Can not read image size from empty metadata");
 }
 
-void ImgSize(const FileName &filename, int &Xdim, int &Ydim, int &Zdim, int &Ndim)
+void ImgSize(const FileName &filename, int &Xdim, int &Ydim, int &Zdim, unsigned long  &Ndim)
 {
     ImgSize(MetaData(filename), Xdim, Ydim, Zdim, Ndim);
 }
